requirements.txt
anthropic>=0.8.0
pandas>=2.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
